PizdeckakMnogo

+++ DataBase/Database/Models/Warns.cs
@@ -4,7 +4,7 @@
     {
         public ulong Id { get; set; }
         public ulong guildid { get; set; }
-        public ulong CountWarn { get; set; }
+        public byte CountWarn { get; set; }
         public string ReportWarn { get; set; }
     }
 }
